move config cmd tests to own module
